chore: temporary fix scalar conflict theme with nextra
