+++ package.json
@@ -26,6 +26,7 @@
     "fs": "^0.0.1-security",
     "gray-matter": "^4.0.3",
     "next": "^14.1.4",
+    "next-seo": "^6.5.0",
     "nextra": "^2.13.4",
     "nextra-theme-docs": "^2.13.4",
     "path": "^0.12.7",
