--- src/pages/api-reference.mdx
@@ -1,3 +1,0 @@
-import APIReference from "@/components/APIReference"
-
-<APIReference />
